+++ renderer/pages/_app.jsx
@@ -9,6 +9,9 @@ import DekAppLayoutWrapper from '@components/core/layout';
 import { CommonAppDataProvider } from '@hooks/useCommonChecks';
 import { LocalizationProvider } from '@hooks/useLocalization';
 import { DeepLinkProvider } from '@hooks/useDeepLinkListener';
+import { ErrorWrapper } from '@components/core/error-wrap';
+import useAppLogger from '@hooks/useAppLogger';
+import { useRouter } from 'next/router';
 
 // Import global stylesheets
 import 'bootstrap/dist/css/bootstrap.css';
@@ -17,13 +20,23 @@ import '@styles/globals.css';
 
 export default function MainAppWrapper({ Component, pageProps }) {
     // console.log({Component, pageProps});
+    const applog = useAppLogger("ErrorWrapper");
+    const router = useRouter();
+
+    const handleError = (error, info) => {
+        applog("error", error.message);
+        applog("error", info.componentStack);
+    };
+
     return <LocalizationProvider>
-        <DeepLinkProvider>
-            <CommonAppDataProvider>
-                <DekAppLayoutWrapper>
-                    <Component {...pageProps}/>
-                </DekAppLayoutWrapper>
-            </CommonAppDataProvider>
-        </DeepLinkProvider>
-    </LocalizationProvider>
+            <DeepLinkProvider>
+                <CommonAppDataProvider>
+                    <DekAppLayoutWrapper>
+                        <ErrorWrapper key={router.pathname} onError={handleError}>
+                            <Component {...pageProps}/>
+                        </ErrorWrapper>
+                    </DekAppLayoutWrapper>
+                </CommonAppDataProvider>
+            </DeepLinkProvider>
+        </LocalizationProvider>
 };